Prefer file extension over content headers
When determining content type the current behaviour is to trust the
`Content-Type` header. However, we have recieved a number of documents
with the wrong value set for this.

File extensions are arguably slightly more trustworthy then the content
type header, so we are opting to prefer it when one matches a supported
file type