Vendorise pdfs for integration test
This removes an external dependency from integration tests that was
causing a failure and blocking deploys. Previously, tests ran against
live pdf files on various government websites. So when those websites
changed the tests no longer worked.

This was blocking other work, so in order to unblock efficiently, I've
simply switched the tests to point at our own staging cdn. This is still
not ideal as now the tests depend on our cdn. But it is an improvement
and means we can progress other work