handle .doc extensions as well as .docx (#131)
## In this PR

- Makes the ingest stage work for doc files (it only worked for docx
before).
- Also intentionally separates out where we define content types from
the SDK, which deals with parser outputs. (I think this is ok as once we
start converting to PDF, we expect the list of content types we handle
in ingest to be a superset of those we handle in the SDK – so we don't
want to bump the SDK for each new content type we handle.)
- Driveby fix for another test that was failing locally - detecting
content type for URLs which don't have a file extension in them (e.g.
htmls). The behaviour now seems to reflect what _should_ happen; I'm not
sure how this failing local test passed before!

## How this is tested

- Adds a test for doc and docx files in the upload files function