download from source: add more request headers & increase timeout (#134)
After the large run of PDFs, all that were valid succeeded except for a
few. Looking at the logs, these seemed to be either due to timeout
issues or 404s. The 404s are strange as the URLs work with a bare
`requests.get` on my laptop – maybe these are websites blocking our
servers?

In this PR I've added a few more browser-y request headers (while
keeping the responsible `User-Agent` one, for now) and bumped the
timeout