fix: prioritise file content to detect MIME type
Our files come from third parties. Third party servers can be unpredictable.
Making the file itself the first port of call for determining file type
means we can more reliabley determine if a file is a pdf or not.

This still falls back to the old strategy, as the library used doesnt
work for html files — and it doesnt hurt to have a fallback.